refactor: move dto to Shared project
