+++ Recademy.Api/Services/Abstraction/IAchievementService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Recademy.Core.Dto;
+using Recademy.Shared.Dtos;
 using Recademy.Core.Models;
 
 namespace Recademy.Api.Services.Abstraction
