+++ Recademy.Api/Services/Abstraction/IReviewResponseService.cs
@@ -1,4 +1,4 @@
-﻿using Recademy.Core.Dto;
+﻿using Recademy.Shared.Dtos;
 
 namespace Recademy.Api.Services.Abstraction
 {
