+++ Recademy.Api/Services/Abstraction/IUserService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Recademy.Core.Dto;
+using Recademy.Shared.Dtos;
 using Recademy.Core.Types;
 
 namespace Recademy.Api.Services.Abstraction
