+++ Recademy.Api/Services/Implementations/ProjectService.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Recademy.Api.Repositories;
 using Recademy.Api.Services.Abstraction;
-using Recademy.Core.Dto;
+using Recademy.Shared.Dtos;
 using Recademy.Core.Models;
 using Recademy.Core.Tools;
