+++ Recademy.Api/Services/Implementations/ReviewService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Recademy.Api.Repositories;
 using Recademy.Api.Services.Abstraction;
-using Recademy.Core.Dto;
+using Recademy.Shared.Dtos;
 using Recademy.Core.Models;
 using Recademy.Core.Tools;
 using Recademy.Core.Types;
