+++ Recademy.Shared/Dtos/GitHubIssueCreateDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Recademy.Core.Dto
+namespace Recademy.Shared.Dtos
 {
     public class GitHubIssueCreateDto
     {
