+++ Recademy.Shared/Dtos/ReviewRequestInfoDto.cs
@@ -1,9 +1,8 @@
-﻿using System;
-using Recademy.Core.Models;
+﻿using Recademy.Core.Models;
 using Recademy.Core.Tools;
 using Recademy.Core.Types;
 
-namespace Recademy.Core.Dto
+namespace Recademy.Shared.Dtos
 {
     public class ReviewRequestInfoDto
     {
