+++ Recademy.Shared/Dtos/TagProfileDto.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Generic;
-
-namespace Recademy.Core.Dto
+﻿namespace Recademy.Shared.Dtos
 {
     public class TagProfileDto
     {
