+++ Recademy.WebUi/Pages/Achievements.razor
@@ -1,5 +1,5 @@
 ﻿@page "/achievements"
-@using Recademy.Core.Dto
+@using Recademy.Shared.Dtos;
 @inject HttpClient Http
 
 <h1>Достижения</h1>
