+++ Recademy.WebUi/Shared/MainLayout.razor
@@ -1,4 +1,4 @@
-﻿@using Recademy.Core.Dto
+﻿@using Recademy.Shared.Dtos;
 @inherits LayoutComponentBase
 @inject ILogger<MainLayout> Logger
 @inject HttpClient Http
